+++ MedHeadBackEnd/src/test/java/com/medhead/emergency/service/TravelTimeCalculatorTest.java
@@ -37,7 +37,7 @@ public class TravelTimeCalculatorTest {
 
     @Test
     void calculateTravelTimeBetweenTwoPointsTest() {
-        int travelTime = travelTimeCalculator.calculateTravelTimeBetweenTwoPoints(testPosition, medicalCenter3.getGeographicCoordinates());
+        long travelTime = travelTimeCalculator.calculateTravelTimeBetweenTwoPoints(testPosition, medicalCenter3.getGeographicCoordinates());
         // Estimated time by Google Maps = 12 minutes => Test if travel time is between 10 and 15 min
         assertTrue(600000 < travelTime);
         System.out.println(travelTime);
